+++ core/Capture.lua
@@ -229,11 +229,22 @@ local function LoginEventHandler()
 	TakeScreenshot()
 end
 
-local function LevelUpEventHandler(level)
+local function LevelUpEventHandler(level, timePlayedOnPreviousLevel)
+	local message
+
 	if AWL.GAME_TYPE_VANILLA or AWL.GAME_TYPE_TBC then
-		Utils:PrintMessage(L["chat.event.level-up.classic.new"]:format(tostring(level)))
+		message = L["chat.event.level-up.classic.new"]:format(tostring(level))
 	elseif AWL.GAME_TYPE_MISTS or AWL.GAME_TYPE_MAINLINE then
-		Utils:PrintMessage(L["chat.event.level-up.retail.new"]:format(tostring(level)))
+		message = L["chat.event.level-up.retail.new"]:format(tostring(level))
+	end
+
+	if message and MEM.Settings.event["level-up-time-played"] and timePlayedOnPreviousLevel then
+		local days, hours, minutes, seconds = Utils:GetDurationParts(timePlayedOnPreviousLevel)
+		message = message .. " - " .. L["chat.event.level-up.time-played"]:format(level - 1, days, hours, minutes, seconds)
+	end
+
+	if message then
+		Utils:PrintMessage(message)
 	end
 
 	TakeScreenshot()