+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Memento/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Memento/releases)
 
-- Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility
+- Added: Level time chat notifications - The time played on the current level can now be shown for all screenshot events
+- Added: Level-up time chat notifications - The time played on the previous level can now be shown in level-up screenshot notifications